+++ lib/json_pointer/utils.ex
@@ -9,9 +9,9 @@ defmodule JSONPointer.Utils do
 
   ## Examples
 
-      iex> JSONPointer.escape "hello~bla"
+      iex> JSONPointer.Utils.escape "hello~bla"
       "hello~0bla"
-      iex> JSONPointer.escape "hello/bla"
+      iex> JSONPointer.Utils.escape "hello/bla"
       "hello~1bla"
 
   """
@@ -28,10 +28,12 @@ defmodule JSONPointer.Utils do
 
   ## Examples
 
-      iex> JSONPointer.unescape "hello~0bla"
+      iex> JSONPointer.Utils.unescape "hello~0bla"
       "hello~bla"
-      iex> JSONPointer.unescape "hello~1bla"
+      iex> JSONPointer.Utils.unescape "hello~1bla"
       "hello/bla"
+      iex> JSONPointer.Utils.unescape "/~01"
+      "/~1"
   """
   @spec unescape(String.t()) :: String.t()
   def unescape(str) do
@@ -45,7 +47,7 @@ defmodule JSONPointer.Utils do
   Converts a JSON pointer into a list of reference tokens
 
   ## Examples
-      iex> JSONPointer.parse("/fridge/butter")
+      iex> JSONPointer.Utils.parse("/fridge/butter")
       {:ok, [ "fridge", "butter"] }
   """
   def parse(""), do: {:ok, []}
@@ -93,11 +95,11 @@ defmodule JSONPointer.Utils do
   Attempts to converts a value into an integer
 
   ## Examples
-      iex> parse_index 10
+      iex> JSONPointer.Utils.parse_index 10
       10
-      iex> parse_index "100"
+      iex> JSONPointer.Utils.parse_index "100"
       100
-      iex> parse_index "92.4"
+      iex> JSONPointer.Utils.parse_index "92.4"
       {:error, "invalid index: 92.4"}
   """
   def parse_index("0"), do: 0
@@ -114,4 +116,38 @@ defmodule JSONPointer.Utils do
         {:error, "invalid index: #{val}"}
     end
   end
+
+  @doc """
+  Checks whether two values are equal, returns an error message if they
+  are not
+
+  ## Examples
+      iex> JSONPointer.Utils.are_equal? "hello", "hello"
+      {:ok, true}
+      iex> JSONPointer.Utils.are_equal? "10", 10
+      {:error, "number is not equal to string"}
+      iex> JSONPointer.Utils.are_equal? [10, "12", 15], [15, "12", 10]
+      {:error, "not equal"}
+  """
+  @spec are_equal?(String.t(), number) :: {:error, String.t()}
+  def are_equal?(val1, val2) when is_binary(val1) and is_number(val2),
+    do: {:error, "number is not equal to string"}
+
+  @spec are_equal?(number, String.t()) :: {:error, String.t()}
+  def are_equal?(val1, val2) when is_number(val1) and is_binary(val2),
+    do: {:error, "string is not equal to number"}
+
+  @spec are_equal?(any(), any()) :: {:ok, true} | {:error, String.t()}
+  def are_equal?(val1, val2) do
+    if val1 == val2 do
+      {:ok, true}
+    else
+      are_not_equal_error(val1, val2)
+    end
+  end
+
+  defp are_not_equal_error(val1, val2) when is_binary(val1) and is_binary(val2),
+    do: {:error, "string not equivalent"}
+
+  defp are_not_equal_error(_val1, _val2), do: {:error, "not equal"}
 end